// sdk/typescript/src/api/artifact/artifact.ts
export enum ArtifactSystem {
  UNKNOWN_SYSTEM = 0,
  AARCH64_DARWIN = 1,
  AARCH64_LINUX = 2,
  X8664_DARWIN = 3,
  X8664_LINUX = 4,
  UNRECOGNIZED = -1,
}